+++ packages/rhf-mui-components/src/mui/select/index.tsx
@@ -31,7 +31,6 @@ type SelectInputProps = Omit<SelectProps,
 | 'value'
 | 'defaultValue'
 | 'onBlur'
-| 'disabled'
 | 'ref'
 >;