+++ packages/rhf-mui-components/src/types/mui.ts
@@ -16,6 +16,7 @@ export type FormControlLabelProps = Omit<
   | 'value'
   | 'defaultValue'
   | 'defaultChecked'
+  | 'disabled'
   | 'key'
 >;